# src/DotNet.IoC.Template.Presentation.WebAPI.Client/openapi-spec.json
{
  "x-generator": "NSwag v13.20.0.0 (NJsonSchema v10.9.0.0 (Newtonsoft.Json v10.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7247"
    }
  ],
  "paths": {
    "/api/Country/{id}": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/Country": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_Post",
        "requestBody": {
          "x-name": "country",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryViewModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_Put",
        "requestBody": {
          "x-name": "updatedCountry",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/Player/{id}": {
      "get": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/Player": {
      "get": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Post",
        "requestBody": {
          "x-name": "player",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerViewModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Put",
        "requestBody": {
          "x-name": "updatedPlayer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/Team/{id}": {
      "get": {
        "tags": [
          "Team"
        ],
        "operationId": "Team_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamViewModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Team"
        ],
        "operationId": "Team_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/Team": {
      "get": {
        "tags": [
          "Team"
        ],
        "operationId": "Team_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Team"
        ],
        "operationId": "Team_Post",
        "requestBody": {
          "x-name": "team",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamViewModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Team"
        ],
        "operationId": "Team_Put",
        "requestBody": {
          "x-name": "updatedTeam",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamViewModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "WeatherForecast_Get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CountryViewModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "isoCode"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "isoCode": {
            "type": "string",
            "maxLength": 10,
            "minLength": 0
          },
          "observations": {
            "type": "string",
            "maxLength": 2040,
            "minLength": 0
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PlayerViewModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "observations": {
            "type": "string",
            "maxLength": 2040,
            "minLength": 0
          },
          "countryId": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "teamId": {
            "type": "string"
          },
          "team": {
            "type": "string"
          }
        }
      },
      "TeamViewModel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "observations": {
            "type": "string",
            "maxLength": 2040,
            "minLength": 0
          },
          "countryId": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "playerIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "WeatherForecast": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}